requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==5.0.2
django-bootstrap4==24.1
django-chunked-upload==2.0.0
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2024.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1